+++ src/main/java/cz/cvut/fsv/webgama/dao/jdbc/JdbcAuthorityDao.java
@@ -44,7 +44,7 @@ public class JdbcAuthorityDao extends JdbcDaoSupport implements AuthorityDao {
 		public Authority mapRow(ResultSet rs, int rowNum) throws SQLException {
 
 			Authority authority = new Authority();
-			authority.setId(Integer.valueOf(rs.getInt("authority_id")));
+			authority.setId(Long.valueOf(rs.getLong("authority_id")));
 			authority.setRole(roleDao.findRoleById(rs.getInt("role_id")));
 			authority.setUser(userDao.findUserById(rs.getInt("user_id")));
